# basyx.submodelregistry/basyx.submodelregistry-service/src/generated/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: info@idtwin.org
    name: Industrial Digital Twin Association (IDTA)
  description: "The Full Profile of the Submodel Registry Service Specification as\
    \ part of the [Specification of the Asset Administration Shell: Part 2](http://industrialdigitaltwin.org/en/content-hub).\
    \  \nPublisher: Industrial Digital Twin Association (IDTA) April 2023"
  license:
    name: CC BY 4.0
    url: https://creativecommons.org/licenses/by/4.0/
  title: DotAAS Part 2 | HTTP/REST | Submodel Registry Service Specification
  version: V3.0.1_SSP-001
servers:
- url: /
paths:
...
  /description:
    get:
      operationId: GetDescription
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceDescription'
          description: Requested Description
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
          description: Forbidden
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
          description: Default error handling for unmentioned status codes
      summary: Returns the self-describing information of a network resource (ServiceDescription)
      tags:
      - Submodel Registry
      x-semanticIds:
      - https://admin-shell.io/aas/API/Descriptor/GetDescription/3/0
      x-accepts: application/json
      x-tags:
      - tag: Submodel Registry
components:
...
  schemas:
...
    Result:
      properties:
        messages:
          items:
            $ref: '#/components/schemas/Message'
          type: array
      type: object
    Message:
      properties:
        code:
          maxLength: 32
          minLength: 1
          type: string
        correlationId:
          maxLength: 128
          minLength: 1
          type: string
        messageType:
          enum:
          - Undefined
          - Info
          - Warning
          - Error
          - Exception
          type: string
        text:
          type: string
        timestamp:
          format: date-time
          type: string
      type: object
    ServiceDescription:
      description: "The Description object enables servers to present their capabilities\
        \ to the clients, in particular which profiles they implement. At least one\
        \ defined profile is required. Additional, proprietary attributes might be\
        \ included. Nevertheless, the server must not expect that a regular client\
        \ understands them."
      example:
        profiles:
        - https://admin-shell.io/aas/API/3/0/AssetAdministrationShellServiceSpecification/SSP-001
        - https://admin-shell.io/aas/API/3/0/AssetAdministrationShellServiceSpecification/SSP-001
      properties:
        profiles:
          items:
            enum:
            - https://admin-shell.io/aas/API/3/0/AssetAdministrationShellServiceSpecification/SSP-001
            - https://admin-shell.io/aas/API/3/0/AssetAdministrationShellServiceSpecification/SSP-002
            - https://admin-shell.io/aas/API/3/0/SubmodelServiceSpecification/SSP-001
            - https://admin-shell.io/aas/API/3/0/SubmodelServiceSpecification/SSP-002
            - https://admin-shell.io/aas/API/3/0/SubmodelServiceSpecification/SSP-003
            - https://admin-shell.io/aas/API/3/0/AasxFileServerServiceSpecification/SSP-001
            - https://admin-shell.io/aas/API/3/0/AssetAdministrationShellRegistryServiceSpecification/SSP-001
            - https://admin-shell.io/aas/API/3/0/AssetAdministrationShellRegistryServiceSpecification/SSP-002
            - https://admin-shell.io/aas/API/3/0/SubmodelRegistryServiceSpecification/SSP-001
            - https://admin-shell.io/aas/API/3/0/SubmodelRegistryServiceSpecification/SSP-002
            - https://admin-shell.io/aas/API/3/0/DiscoveryServiceSpecification/SSP-001
            - https://admin-shell.io/aas/API/3/0/AssetAdministrationShellRepositoryServiceSpecification/SSP-001
            - https://admin-shell.io/aas/API/3/0/AssetAdministrationShellRepositoryServiceSpecification/SSP-002
            - https://admin-shell.io/aas/API/3/0/SubmodelRepositoryServiceSpecification/SSP-001
            - https://admin-shell.io/aas/API/3/0/SubmodelRepositoryServiceSpecification/SSP-002
            - https://admin-shell.io/aas/API/3/0/SubmodelRepositoryServiceSpecification/SSP-003
            - https://admin-shell.io/aas/API/3/0/SubmodelRepositoryServiceSpecification/SSP-004
            - https://admin-shell.io/aas/API/3/0/ConceptDescriptionServiceSpecification/SSP-001
            type: string
          minItems: 1
          type: array
      type: object
...